Done with login and Home screen

+++ Qnopy_Task/Products.swift
@@ -4,13 +4,12 @@ struct Roots:Decodable {
     var products: [Product]
     
 }
-
 struct Product:Decodable {
     var id: Int
     var title: String
     var description: String
     var price: Int
-    var discountPercentage:Float64
+    var discountPercentage:Int
     var rating: Double
     var stock: Int
     var brand: String
